rename : presenter, contract 패키지 수정
