+++ app/src/main/java/woowacourse/movie/view/booking/BookingActivity.kt
@@ -20,8 +20,6 @@ import woowacourse.movie.data.MovieStore
 import woowacourse.movie.domain.model.booking.Booking
 import woowacourse.movie.domain.model.booking.PeopleCount
 import woowacourse.movie.domain.model.movies.Movie
-import woowacourse.movie.presenter.booking.BookingContract
-import woowacourse.movie.presenter.booking.BookingPresenter
 import woowacourse.movie.view.StringFormatter
 import woowacourse.movie.view.ext.toDrawableResourceId
 import woowacourse.movie.view.seat.SeatActivity
