+++ app/src/main/java/woowacourse/movie/view/seat/SeatActivity.kt
@@ -16,8 +16,6 @@ import woowacourse.movie.domain.model.booking.Booking
 import woowacourse.movie.domain.model.seat.Seat
 import woowacourse.movie.domain.model.seat.Seats
 import woowacourse.movie.domain.model.ticket.Ticket
-import woowacourse.movie.presenter.seat.SeatContract
-import woowacourse.movie.presenter.seat.SeatPresenter
 import woowacourse.movie.view.StringFormatter
 import woowacourse.movie.view.complete.BookingCompleteActivity
 import woowacourse.movie.view.ext.getSerializable
